third_party/WebKit/LayoutTests/fast/events/relatedevent.html

<!DOCTYPE html>
<html>
<head>
<script src="../../resources/js-test.js"></script>
</head>
<script>

description("Tests to confirm behavior of updating IDL attributes of RelatedEvent.");

var relatedEvent = new RelatedEvent("foo");

// We should not be able to update readonly attributes.
relatedEvent.type = "bar";
relatedEvent.bubbles = true;
relatedEvent.cancelable = true;
relatedEvent.relatedTarget = document;

shouldBeEqualToString("relatedEvent.type", "foo");
shouldBeFalse("relatedEvent.bubbles");
shouldBeFalse("relatedEvent.cancelable");
shouldBeNull("relatedEvent.relatedTarget");
</script>
</html>
