third_party/WebKit/LayoutTests/fast/events/relatedevent-expected.txt

Tests to confirm behavior of updating IDL attributes of RelatedEvent.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS relatedEvent.type is "foo"
PASS relatedEvent.bubbles is false
PASS relatedEvent.cancelable is false
PASS relatedEvent.relatedTarget is null
PASS successfullyParsed is true

TEST COMPLETE
